Scale fix and continuation of fast controller handling
Scale back to 1
Collision behavior alone will not be enough to catch fast controller.
Experimenting with Raycast or using Trigger status instead. Trigger is
promising as it is physx independent and increments at all speeds, but
determining hit location and applying force is going to be interesting.
Raycast comes with hit location, but is subject to a lesser degree to
the collision issue. Mainly that the controller, aided by the long lever
of the sword, can get entirely through the dummy shinai in less than
180th sec.